refactor(auth): inline security utils in routes
